requirements.txt
streamlit
llama_index
qdrant_client
python-dotenv
pypdf
transformers
streamlit-chat
openai
loguru
streamlit_chat